+++ packages/wiz-ui-react/src/components/base/navigation/components/navigation-container.tsx
@@ -20,14 +20,14 @@ const NavigationContainer: FC<Props> = ({
   className,
   style,
   isOpen,
-  width,
+  width = "180px",
   sticky = false,
   children,
   footer,
 }) => {
-  const widthStyle =
-    width ??
-    (isOpen ? "180px" : `calc(${THEME.spacing.xl} * 2 + ${THEME.spacing.sm})`);
+  const widthStyle = isOpen
+    ? width
+    : `calc(${THEME.spacing.xl} * 2 + ${THEME.spacing.sm})`;
 
   const stickyStyle: CSSProperties = sticky
     ? {